fix loading variable location
the `loading` variable should be defined in `this.state` but was mistakenly
defined in `this.query`, making it by default always rendered as `false` when
query. this makes the datatable thinks the loading is finished while it's still loading the data.

this fixes issue https://github.com/CAIDA/bgphijacks-tools/issues/425