fix defcon pfx events table rendering error
The issue is that for some defcon cases, the origin is the newcomer, and for
some it's the oldcomer. The previous version of the dashboard js code assumes
oldcomer exists, while not every case is so. We fixed it by allowing both
newcomer and oldcomer to be associated with the prefix in the prefix events
table.